services/api/requirements.txt
# TuneTrail API Service - Python Dependencies
# Updated: September 2025

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.4

# Redis & Caching
redis[hiredis]==5.1.1
hiredis==3.0.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.2.0

# ML & AI
torch==2.5.1
numpy==2.1.2
scikit-learn==1.5.2

# Audio Processing
librosa==0.10.2.post1
soundfile==0.12.1

# Task Queue
celery[redis]==5.4.0
kombu==5.4.2

# HTTP & API
httpx==0.27.2
aiofiles==24.1.0

# Storage
boto3==1.35.34
minio==7.2.9

# Monitoring & Logging
sentry-sdk[fastapi]==2.14.0
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==7.0.0

# Utilities
python-dotenv==1.0.1
pytz==2024.2
email-validator==2.2.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.8.0
ruff==0.6.9
mypy==1.11.2